Moved conversions out and added tests
